Refactor: seperate utils
